Move vulkan resources into new module
