+++ src/lib.rs
@@ -1,5 +1,6 @@
 //! Radiosity
 
 mod associations;
+pub mod vulkan;
 
 pub use associations::Associated;
